item-manager-app/backend/requirements.txt
flask
gunicorn
psycopg2-binary
python-dotenv
flask-cors
